+++ js/models/event_handler.js
@@ -129,13 +129,6 @@ export default class EventHandler {
       this.aladin.setFoV(fov);
     });
 
-    this.aladin.on("layerChanged", (imageLayer, layerName, state) => {
-      if (layerName !== "base" || state !== "ADDED") return;
-      this.updateWCS();
-      this.model.set("_base_layer_last_view", imageLayer.id);
-      this.model.save_changes();
-    });
-
     /* Div control */
     this.model.on("change:_height", () => {
       let height = this.model.get("_height");
@@ -155,6 +148,15 @@ export default class EventHandler {
 
     this.aladin.on("projectionChanged", () => {
       this.updateWCS();
+      this.model.save_changes();
+    });
+
+    this.aladin.on("layerChanged", (imageLayer, layerName, state) => {
+      if (layerName === "base")
+        this.model.set("_survey_body", imageLayer.hipsBody || "sky");
+      if (layerName !== "base" || state !== "ADDED") return;
+      this.updateWCS();
+      this.model.set("_base_layer_last_view", imageLayer.id);
       this.model.save_changes();
     });