+++ src/ipyaladin/__about__.py
@@ -1,2 +1,2 @@
 __version__ = "0.4.0"
-__aladin_lite_version__ = "3.4.4-beta"
+__aladin_lite_version__ = "3.5.1-alpha"