+++ src/ipyaladin/utils/_coordinate_parser.py
@@ -1,16 +1,25 @@
+import warnings
 from typing import Tuple
 
-from astropy.coordinates import SkyCoord, Angle
+import requests
+from astropy.coordinates import SkyCoord, Angle, Longitude, Latitude
 import re
 
+from ipyaladin.utils.exceptions import NameResolverWarning
 
-def parse_coordinate_string(string: str) -> SkyCoord:
+
+def _parse_coordinate_string(
+    string: str, body: str = "sky"
+) -> Tuple[Longitude, Latitude]:
     """Parse a string containing coordinates.
 
     Parameters
     ----------
     string : str
         The string containing the coordinates.
+    body : str
+        The planetary body to use for the coordinates. Default
+        is "sky" when there is no planetary body.
 
     Returns
     -------
@@ -19,20 +28,71 @@ def parse_coordinate_string(string: str) -> SkyCoord:
 
     """
     if not _is_coordinate_string(string):
-        return SkyCoord.from_name(string)
-    coordinates: Tuple[str, str] = _split_coordinate_string(string)
+        if body == "sky" or body is None:
+            sesame = SkyCoord.from_name(string)
+            return sesame.icrs.ra.deg, sesame.icrs.dec.deg
+        return _from_name_on_planet(string, body)
+    # coordinates should be parsed from the string
+    coordinates = _split_coordinate_string(string)
     # Parse ra and dec to astropy Angle objects
-    dec: Angle = Angle(coordinates[1], unit="deg")
+    lat: Angle = Angle(coordinates[1], unit="deg")
     if _is_hour_angle_string(coordinates[0]):
-        ra = Angle(coordinates[0], unit="hour")
+        lon = Angle(coordinates[0], unit="hour")
     else:
-        ra = Angle(coordinates[0], unit="deg")
+        lon = Angle(coordinates[0], unit="deg")
     # Create SkyCoord object
     if string[0] == "B":
-        return SkyCoord(ra=ra, dec=dec, equinox="B1950", frame="fk4")
-    if string[0] == "G":
-        return SkyCoord(l=ra, b=dec, frame="galactic")
-    return SkyCoord(ra=ra, dec=dec, frame="icrs")
+        coo = SkyCoord(ra=lon, dec=lat, equinox="B1950", frame="fk4")
+    elif string[0] == "G":
+        coo = SkyCoord(l=lon, b=lat, frame="galactic")
+    else:
+        coo = SkyCoord(ra=lon, dec=lat, frame="icrs")
+    return coo.icrs.ra.deg, coo.icrs.dec.deg
+
+
+def _from_name_on_planet(string: str, body: str) -> SkyCoord:
+    """Get coordinates from a name on a planetary body.
+
+    Parameters
+    ----------
+    string : str
+        The name of the feature.
+    body : str
+        The planetary body to use for the coordinates.
+
+    Returns
+    -------
+    SkyCoord
+        An `astropy.coordinates.SkyCoord` object representing the coordinates.
+    """
+    url = (
+        f"https://alasky.cds.unistra.fr/planetary-features/resolve"
+        f"?identifier={string.replace(' ', '+')}&body={body}&threshold=0.7&format=json"
+    )
+    request = requests.get(url)
+    if request.status_code != requests.codes.ok:
+        raise ValueError(
+            "No coordinates found for this requested planetary feature: " f"'{string}'"
+        )
+    data = request.json()["data"]
+    # response is different for earth
+    if body == "earth":
+        return data[0][0], data[0][1]
+    # case of every other planetary bodies
+    identifier = data[0][1]
+    lat = data[0][5]  # inverted lon and lat in response
+    lon = data[0][6]
+    system = data[0][11]
+    if identifier != string:
+        warnings.warn(
+            f"Nothing found for '{string}' on {body}. However, a {identifier} exists. "
+            f"Moving to {identifier}.",
+            NameResolverWarning,
+            stacklevel=2,
+        )
+    if "+West" in system:
+        lon = 360 - lon
+    return lon, lat
 
 
 def _is_coordinate_string(string: str) -> bool: