+++ src/ipyaladin/utils/exceptions.py
@@ -12,3 +12,11 @@ class WidgetNotReadyError(OSError):
     def __init__(self, message: str) -> None:
         self.message = message
         super(WidgetNotReadyError, self).__init__(message)
+
+
+class NameResolverWarning(UserWarning):
+    """Warning raised when a name resolver is used."""
+
+    def __init__(self, message: str) -> None:
+        self.message = message
+        super(NameResolverWarning, self).__init__(message)