+++ src/ipyaladin/widget.py
@@ -15,7 +15,8 @@ from typing import ClassVar, Dict, Final, List, Optional, Tuple, Union
 import warnings
 
 import anywidget
-from astropy.coordinates import SkyCoord, Angle
+from astropy.coordinates import SkyCoord, Angle, Longitude, Latitude
+from astropy.coordinates.name_resolve import NameResolveError
 from astropy.table.table import QTable
 from astropy.table import Table
 from astropy.io import fits as astropy_fits
@@ -25,7 +26,7 @@ import numpy as np
 import traitlets
 
 from .utils.exceptions import WidgetReducedError, WidgetNotReadyError
-from .utils._coordinate_parser import parse_coordinate_string
+from .utils._coordinate_parser import _parse_coordinate_string
 
 try:
     from regions import (
@@ -237,6 +238,10 @@ class Aladin(anywidget.AnyWidget):
     )
 
     # Values
+    _ready = Bool(
+        False,
+        help="A private trait that stores the readiness of the widget.",
+    ).tag(sync=True)
     _wcs = traitlets.Dict().tag(sync=True)
     _fov_xy = traitlets.Dict().tag(sync=True)
 
@@ -250,6 +255,10 @@ class Aladin(anywidget.AnyWidget):
     listener_callback: ClassVar[Dict[str, callable]] = {}
 
     # overlay survey
+    _survey_body = Unicode(
+        "sky",
+        help="The body name of the base layer survey, 'sky' for the sky survey",
+    ).tag(sync=True, init_option=True)
     overlay_survey = Unicode("").tag(sync=True, init_option=True)
     overlay_survey_opacity = Float(0.0).tag(sync=True, init_option=True)
     _base_layer_last_view = Unicode(
@@ -422,40 +431,79 @@ class Aladin(anywidget.AnyWidget):
         self._wcs = {}
 
     @property
-    def target(self) -> SkyCoord:
+    def target(self) -> Union[SkyCoord, Tuple[float, float]]:
         """The target of the Aladin Lite widget.
 
-        It can be set with either a string or a `~astropy.coordinates.SkyCoord` object.
+        The target can be provided as coordinates (either
+        `~astropy.coordinates.SkyCoord` or (`~astropy.coordinates.Longitude`,
+        `~astropy.coordinates.Latitude`)) or as a name (as a string).
+
+        The conversion from a name to coordinates does call different online services
+        depending on the base layer of the widget:
+        - for the sky: it uses CDS's
+        `Sesame <https://vizier.cds.unistra.fr/viz-bin/Sesame>`_  name resolver. This
+        looks first in SIMBAD, then in VizieR, and then in NED for the object's
+        coordinates.
+        - for Earth: it uses `Nominatim
+        <https://nominatim.openstreetmap.org/ui/search.html>`_, OpenStreetMap's
+        search bar
+        - for other solar system bodies: it calls the `Gazetteer of Planetary
+        Nomenclature <https://planetarynames.wr.usgs.gov/Nomenclature>`_
 
         Returns
         -------
-        `~astropy.coordinates.SkyCoord`
-            A `~astropy.coordinates.SkyCoord` object representing the target.
+        `~astropy.coordinates.SkyCoord` or Tuple[Longitude, Latitude]
+            If ipyaladin has a sky survey, returns a `~astropy.coordinates.SkyCoord`
+            object. Otherwise, if ipyaladin is currently on a planetary body, returns
+            a tuple with the Longitude and the Latitude of the current position of the
+            widget.
 
         """
-        ra, dec = self._target.split(" ")
-        return SkyCoord(
-            ra=ra,
-            dec=dec,
-            frame="icrs",
-            unit="deg",
-        )
+        lon, lat = self._target.split(" ")
+        lon, lat = float(lon), float(lat)
+        if self._survey_body == "sky":
+            return SkyCoord(
+                ra=lon,
+                dec=lat,
+                frame="icrs",
+                unit="deg",
+            )
+        return Longitude(lon, unit="deg"), Latitude(lat, unit="deg")
 
     @target.setter
-    def target(self, target: Union[str, SkyCoord]) -> None:
-        if isinstance(target, str):  # If the target is str, parse it
-            target = parse_coordinate_string(target)
-        elif not isinstance(target, SkyCoord):  # If the target is not str or SkyCoord
+    def target(self, target: Union[str, SkyCoord, Tuple[float, float]]) -> None:
+        if isinstance(target, Tuple):
+            lon, lat = target[0].deg, target[1].deg
+        elif isinstance(target, str):  # If the target is a string, parse it
+            try:
+                lon, lat = _parse_coordinate_string(target, self._survey_body)
+            except NameResolveError as e:
+                # If the widget is not ready, we don't know if the base survey is
+                # celestial or planetary so the error can be caused by two factors
+                if not self._ready:
+                    raise WidgetNotReadyError(
+                        f"Either '{target}' is not a valid celestial object name, "
+                        f"or the survey body type is not yet defined so you "
+                        f"need to wait for the widget to be fully loaded."
+                    ) from e
+                # If the widget is ready, the error is caused by the target name
+                # that is not a valid celestial object name
+                raise e
+        elif not isinstance(target, SkyCoord) and not isinstance(target, Tuple):
             raise ValueError(
-                "target must be a string or an astropy.coordinates.SkyCoord object"
+                "target must be a string, an astropy.coordinates.SkyCoord "
+                "object or a tuple of two angle-like astropy quantities."
             )
-        self._target = f"{target.icrs.ra.deg} {target.icrs.dec.deg}"
         self._wcs = {}
+        if isinstance(target, SkyCoord):
+            lon, lat = target.icrs.ra.deg, target.icrs.dec.deg
+
+        self._target = f"{lon} {lat}"
         self.send(
             {
                 "event_name": "goto_ra_dec",
-                "ra": target.icrs.ra.deg,
-                "dec": target.icrs.dec.deg,
+                "ra": lon,
+                "dec": lat,
             }
         )