+++ src/test/test_aladin.py
@@ -1,10 +1,12 @@
-from astropy.coordinates import Angle, SkyCoord
+from astropy.coordinates import Angle, SkyCoord, Longitude, Latitude
+import astropy.units as u
 import numpy as np
 import pytest
-from typing import Callable
+import requests
+from typing import Callable, Dict
 
 from ipyaladin import Aladin
-from ipyaladin.utils._coordinate_parser import parse_coordinate_string
+from ipyaladin.utils._coordinate_parser import _parse_coordinate_string
 
 from .test_coordinate_parser import test_is_coordinate_string_values
 
@@ -16,6 +18,40 @@ aladin = Aladin()
 def mock_sesame(monkeypatch: Callable) -> None:
     """Sesame calls mocked."""
     monkeypatch.setattr(SkyCoord, "from_name", lambda _: SkyCoord(0, 0, unit="deg"))
+
+
+class MockResponse:
+    """Mock response object for requests.get."""
+
+    def __init__(self) -> None:
+        self.status_code = 200
+
+    def json(self) -> Dict:
+        """Return a mock JSON response."""
+        return {
+            "data": [
+                [
+                    None,
+                    None,
+                    None,
+                    None,
+                    None,
+                    0,
+                    0,
+                    None,
+                    None,
+                    None,
+                    None,
+                    "",
+                ]
+            ]
+        }
+
+
+@pytest.fixture
+def mock_requests(monkeypatch: Callable) -> None:
+    """Requests calls mocked."""  # noqa: D401
+    monkeypatch.setattr(requests, "get", lambda _: MockResponse())
 
 
 test_aladin_string_target, _ = zip(*test_is_coordinate_string_values)
@@ -31,10 +67,39 @@ def test_aladin_string_target_set(target: str, mock_sesame: Callable) -> None:  
         The target string.
 
     """
+    aladin._survey_body = "sky"
     aladin.target = target
-    parsed_target = parse_coordinate_string(target)
-    assert np.isclose(aladin.target.icrs.ra.deg, parsed_target.icrs.ra.deg)
-    assert np.isclose(aladin.target.icrs.dec.deg, parsed_target.icrs.dec.deg)
+    parsed_target = _parse_coordinate_string(target)
+    assert np.isclose(aladin.target.icrs.ra.deg, parsed_target[0])
+    assert np.isclose(aladin.target.icrs.dec.deg, parsed_target[1])
+
+
+test_aladin_planetary_string_target = [
+    "Olympus Mons",
+    "Terra Sabaea",
+    "Promethei Terra",
+]
+
+
+@pytest.mark.filterwarnings("ignore: Nothing found for")
+@pytest.mark.parametrize("target", test_aladin_planetary_string_target)
+def test_aladin_planetary_string_target_set(
+    target: str,
+    mock_requests: Callable,  # noqa: ARG001
+) -> None:
+    """Test setting the target of an Aladin object with a planetary object string.
+
+    Parameters
+    ----------
+    target : str
+        The target string.
+
+    """
+    aladin._survey_body = "mars"
+    aladin.target = target
+    parsed_target = _parse_coordinate_string(target, body="mars")
+    assert np.isclose(aladin.target[0], parsed_target[0])
+    assert np.isclose(aladin.target[1], parsed_target[1])
 
 
 @pytest.mark.parametrize("target", test_aladin_string_target)
@@ -47,10 +112,14 @@ def test_aladin_sky_coord_target_set(target: str, mock_sesame: Callable) -> None
         The target string.
 
     """
-    sc_target = parse_coordinate_string(target)
-    aladin.target = sc_target
-    assert np.isclose(aladin.target.icrs.ra.deg, sc_target.icrs.ra.deg)
-    assert np.isclose(aladin.target.icrs.dec.deg, sc_target.icrs.dec.deg)
+    aladin._survey_body = "sky"
+    sc_target = _parse_coordinate_string(target)
+    aladin.target = (
+        Longitude(sc_target[0], unit=u.deg),
+        Latitude(sc_target[1], unit=u.deg),
+    )
+    assert np.isclose(aladin.target.icrs.ra.deg, sc_target[0])
+    assert np.isclose(aladin.target.icrs.dec.deg, sc_target[1])
 
 
 test_aladin_float_fov = [